+++ src/Products/PloneMeeting/tests/testMeetingConfig.py
@@ -2687,8 +2687,7 @@ class testMeetingConfig(PloneMeetingTestCase):
         self.assertTrue(template.showMeetingManagerReservedField('notes'))
         self.assertTrue(template.mayQuickEdit('notes'))
         # but it does not have access on a real item
-        self._addPrincipalToGroup(
-            self.member.id, get_plone_group_id(self.developers_creators)
+        self._addPrincipalToGroup(self.member.id, self.developers_creators)
         item = self.create('MeetingItem')
         self.assertTrue(item.attribute_is_used('textCheckList'))
         self.assertFalse(item.showMeetingManagerReservedField('textCheckList'))