+++ CHANGES.rst
@@ -15,10 +15,15 @@ Changelog
 - Added `ToolPloneMeeting.showExternalLinksSection` to be able to enable
   the `External links` collapsible on items on a per `MeetingConfig` basis.
   [gbastien]
-- Added new datagrid field `MeetingConfig.CssTransforms` in replacement for
+- Added new datagrid field `MeetingConfig.cssTransforms` in replacement for
   fields `MeetingConfig.cssClassesToHide` and `MeetingConfig.hideCssClassesTo`,
   new field manage `CSS removal` and `content replacement` based on a CSS class,
   this will be used to anonymize content for selected power observers.
+  [gbastien]
+- Added possibilty to display a `description` when using the `PMCheckBoxWidget`
+  by defining a `description` on the `vocabulary term`.
+  [gbastien]
+- Added new batch action `Insert barcode` on annexes and decision annexes.
   [gbastien]
 
 4.2.29rc8 (2026-04-01)